ui bug fix
