--- OsmAnd/src/net/osmand/plus/osmedit/opr/OPRImage.java
@@ -1,8 +1,0 @@
-package net.osmand.plus.osmedit.opr;
-
-public class OPRImage {
-	public String type;
-	public String hash;
-	public String cid;
-	public String extension;
-}
